Split doctor checks into platform, ASR, and shared type modules.
Keep doctor.py as the orchestration entry point while moving platform
readiness, backend runtime, and Check types into focused modules.
